New test - loading image and rendering texture
- new GL program - TextureProgram, some new classes will be introduced
into gamebase
- test of new system of interfaces (IPositionable, IDrawable, etc.)
- built ImageMagick (library for loading images)
- changed structure of repository: new directories package and resources
- some fixes in gamebase

+++ src/gamebase/include/gamebase/utils/Size.h
@@ -4,10 +4,10 @@ namespace gamebase {
 
 struct Size {
     Size() : width(0), height(0) {}
-    Size(float w, float h) : width(w), height(h) {}
+    Size(unsigned int w, unsigned int h) : width(w), height(h) {}
 
-    float width;
-    float height;
+    unsigned int width;
+    unsigned int height;
 };
 
 }
